docs: relocate data-grid filtering examples
