--- docs/app/docs/data-grid/[slug]/components/theme/options.demo.tsx
@@ -1,10 +1,0 @@
-import { type ThemeOptions } from '@cerberus-design/data-grid'
-
-const _customTheme: ThemeOptions = {
-  border: 'none',
-  borderColor: 'transparent',
-  rowBgColor: 'var(--cerberus-colors-page-surface-initial)',
-  rowEvenBgColor: 'var(--cerberus-colors-page-surface-initial)',
-  rowHoverBgColor: 'transparent',
-  headCellBorderBottomColor: 'transparent',
-}
