--- docs/app/docs/data-grid/[slug]/components/theme/static.tsx
@@ -1,13 +1,0 @@
-import { BasicDemo } from './basic.demo'
-
-export const DEMOS = {
-  basic: {
-    id: 'theme.basic',
-    preview: <BasicDemo />,
-    context: 'data-grid',
-  },
-  options: {
-    id: 'theme.options',
-    context: 'data-grid',
-  },
-}
