+++ docs/components/example/index.tsx
@@ -89,13 +89,7 @@ async function ExampleContent({ path, demo }: ExampleProps) {
       rounded="lg"
       w="full"
     >
-      <HStack
-        data-slot="example:preview"
-        contain="content"
-        justify="center"
-        py="md"
-        w="full"
-      >
+      <HStack data-slot="example:preview" justify="center" py="md" w="full">
         {previewNode}
       </HStack>
