+++ docs/content/blog/v1-2-release.mdx
@@ -10,8 +10,6 @@ timeToRead: 2
 tag: Release
 ---
 
-import { DEMOS as GridFilterDemos } from '@/app/docs/data-grid/[slug]/components/filtering/static'
-
 This release adds new features to React and the Data Grid.
 
 ## Overview
@@ -107,7 +105,7 @@ Once a filter is active, the user can:
   }
 />
 
-<CodePreview {...GridFilterDemos.basic} />
+<Example path="data-grid/filtering" demo="basic" />
 <br />
 
 Checkout the [Data Grid Filtering documentation](/docs/data-grid/filtering) for more details.
