+++ docs/examples/data-grid/filtering/compare.demo.tsx
@@ -1,24 +1,8 @@
-'use client'
-
-import { Edit } from '@carbon/icons-react'
-import { Format, IconButton, Text } from '@cerberus-design/react'
-import { HStack, VStack } from 'styled-system/jsx'
-import { columnHelper } from '../quick-start/helper.demo'
+import { VStack } from '@/styled-system/jsx'
+import { Text } from '@cerberus-design/react'
+import { columnHelper } from '../col-defs/employees'
 
 export const columns = [
-  columnHelper.accessor('id', {
-    header: 'ID',
-    width: 80,
-    features: {
-      pinning: {
-        defaultPosition: 'left',
-      },
-      sort: true,
-      filter: true,
-    },
-    cell: ({ value }) => <Text>#{value}</Text>,
-  }),
-
   columnHelper.accessorFn((row) => `${row.firstName} ${row.lastName}`, {
     id: 'fullName',
     header: 'Employee',
@@ -47,34 +31,6 @@ export const columns = [
           {row.email}
         </Text>
       </VStack>
-    ),
-  }),
-
-  columnHelper.accessorFn((row) => row.department.name, {
-    id: 'department',
-    header: 'Department',
-    features: { sort: true, filter: true },
-  }),
-
-  columnHelper.accessor('salary', {
-    header: 'Salary',
-    width: 135,
-    features: { sort: true },
-    cell: ({ value }) => (
-      <Format.Number value={value} style="currency" currency="USD" />
-    ),
-  }),
-
-  columnHelper.display({
-    id: 'actions',
-    header: 'Actions',
-    width: 125,
-    cell: () => (
-      <HStack justify="center" w="full">
-        <IconButton ariaLabel="View more options">
-          <Edit />
-        </IconButton>
-      </HStack>
     ),
   }),
 ]
