+++ docs/examples/data-grid/filtering/global.demo.tsx
