+++ docs/examples/data-grid/filtering/index.tsx
@@ -3,21 +3,15 @@ import { GlobalDemo } from './global.demo'
 
 export const DEMOS = {
   basic: {
-    id: 'filtering.basic',
     preview: <BasicDemo />,
-    context: 'data-grid',
   },
   global: {
-    id: 'filtering.global',
     preview: <GlobalDemo />,
-    context: 'data-grid',
   },
   compare: {
-    id: 'filtering.compare',
-    context: 'data-grid',
+    preview: null,
   },
   operators: {
-    id: 'filtering.operators',
-    context: 'data-grid',
+    preview: null,
   },
 }
